New folders and setup for pakage
