+++ THERMALCAMERA_S3/videomaker.py
@@ -1,0 +1,63 @@
+from datetime import datetime
+import numpy as np
+import cv2
+
+from THERMALCAMERA_S3 import THERMALCAMERA_S3_DATA
+
+class VideoMaker:
+    """
+    Class used to save a video of the plot
+    """
+
+    def __init__(self, size=(720,960), fps=4):
+        self.filming = False
+        self.size = size
+        self.fps = fps
+
+    def start_video(self):
+        """
+        Create a VideoWriter object
+
+        Parameters
+        ----------
+        none
+        """
+
+        now = datetime.now() # current date and time
+        time = now.strftime("%d_%m_%Y__%H_%M_%S")
+        filename = THERMALCAMERA_S3_DATA/f"{time}.mp4"
+        fourcc = cv2.VideoWriter_fourcc(*'mp4v')
+        self.video = cv2.VideoWriter(filename, fourcc, self.fps, self.size, isColor=True)
+        self.filming = True
+        print(f"Filming {filename}")
+
+    def add_frame(self, fig):
+        """
+        Add frame to video if filming, else do nothing
+
+        Parameters
+        ----------
+        fig : figure
+        """
+
+        if self.filming:
+            data_arr = np.asarray(fig.canvas.renderer.buffer_rgba())
+            data_arr = cv2.cvtColor(data_arr, cv2.COLOR_RGBA2BGR) # Convert to BGR (opencv's default)
+            data_arr = cv2.resize(data_arr, self.size) # resize image to video size
+            self.video.write(data_arr) # add image to video writer
+        else:
+            pass
+
+    def stop_video(self):
+        """
+        Save output video
+
+        Parameters
+        ----------
+        none
+        """
+
+        self.video.release()
+        self.filming = False
+        print(f"Stopped filming, saved output video")
+
